requirements.txt
neo4j==5.27.0
python-dotenv==1.0.0
openai==0.28.0
transformers==4.35.2
torch==2.2.0
numpy==1.26.2
langchain==0.0.350
pandas==2.1.3
click==8.1.7
rich==13.7.0
openpyxl==3.1.2
